Pack a MAC address into a u32 without shifting into the sign bit
The four WPA parsers build their PMKID input out of the two MAC addresses a
byte at a time:

    wpa->pmkid_data[4] = (mac_sta[2] <<  0)
                       | (mac_sta[3] <<  8)
                       | (mac_sta[4] << 16)
                       | (mac_sta[5] << 24);

Each byte promotes to int, so a MAC byte of 0x80 or more shifts into the sign
bit of a signed 32 bit type, which is undefined. The compilers hashcat is built
with all produce the value the code wants, and a cast says so rather than
relying on it.

UndefinedBehaviorSanitizer reports it on the example hash itself, before any
fuzzing starts, so the seed corpus of the target is enough:

    CFLAGS="-g -O1 -fsanitize=fuzzer-no-link,address,undefined -fno-omit-frame-pointer" \
      tools/fuzz/build.sh
    UBSAN_OPTIONS=halt_on_error=1 ./fuzz_out/fuzz_parse_22000 fuzz_out/work/seeds/parse_22000

    before: src/modules/module_22000.c:1007:38: runtime error: left shift of 244
            by 24 places cannot be represented in type 'int'
    after:  no report

m22001 has the same three lines, and m16800 and m16801 have three each of their
own, so all four are cast. Every one of them still cracks its own example hash,
the two deprecated ones with --deprecated-check-disable.

+++ src/modules/module_16801.c
@@ -459,15 +459,15 @@ int module_hash_decode (MAYBE_UNUSED const hashconfig_t *hashconfig, MAYBE_UNUSE
   wpa_pmkid->pmkid_data[2] = (wpa_pmkid->orig_mac_ap[0]  <<  0)
                            | (wpa_pmkid->orig_mac_ap[1]  <<  8)
                            | (wpa_pmkid->orig_mac_ap[2]  << 16)
-                           | (wpa_pmkid->orig_mac_ap[3]  << 24);
+                           | ((u32) wpa_pmkid->orig_mac_ap[3] << 24);
   wpa_pmkid->pmkid_data[3] = (wpa_pmkid->orig_mac_ap[4]  <<  0)
                            | (wpa_pmkid->orig_mac_ap[5]  <<  8)
                            | (wpa_pmkid->orig_mac_sta[0] << 16)
-                           | (wpa_pmkid->orig_mac_sta[1] << 24);
+                           | ((u32) wpa_pmkid->orig_mac_sta[1] << 24);
   wpa_pmkid->pmkid_data[4] = (wpa_pmkid->orig_mac_sta[2] <<  0)
                            | (wpa_pmkid->orig_mac_sta[3] <<  8)
                            | (wpa_pmkid->orig_mac_sta[4] << 16)
-                           | (wpa_pmkid->orig_mac_sta[5] << 24);
+                           | ((u32) wpa_pmkid->orig_mac_sta[5] << 24);
 
   // salt
